refactor: extract user domain to repository/user.rs
