+++ backend-v2/src/repository/mod.rs
@@ -1,4 +1,6 @@
 mod sqlite;
+mod user;
+pub use user::{DbUser, NewUser, UpdateUser};
 
 use std::{sync::Arc, time::Duration};
 
@@ -9,8 +11,6 @@ use serde::{Deserialize, Serialize};
 use sqlx::sqlite::{SqliteConnectOptions, SqliteJournalMode, SqlitePoolOptions, SqliteSynchronous};
 use thiserror::Error;
 use uuid::Uuid;
-
-use crate::auth::Role;
 
 const DEFAULT_DB_PATH: &str = "sqlite://./data/agent-k.db";
 
@@ -96,34 +96,6 @@ pub struct DbSession {
     pub share_mode: ShareMode,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
-}
-
-#[derive(Debug, Clone)]
-pub struct DbUser {
-    pub id: Uuid,
-    pub username: String,
-    pub password_hash: String,
-    pub role: Role,
-    pub display_name: Option<String>,
-    pub is_active: bool,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-}
-
-pub struct NewUser {
-    pub id: Uuid,
-    pub username: String,
-    pub password_hash: String,
-    pub role: Role,
-    pub display_name: Option<String>,
-    pub is_active: bool,
-}
-
-pub struct UpdateUser {
-    pub display_name: Option<String>,
-    pub password_hash: Option<String>,
-    pub role: Option<Role>,
-    pub is_active: Option<bool>,
 }
 
 pub type AppRepository = Arc<SqliteRepository>;
